full/users_analysis/requirements.txt
pyspark
python-dateutil
numpy
pandas
scikit-learn
torch
networkx
confluent_kafka
requests
